+++ service/controller/cluster.go
@@ -10,11 +10,12 @@ import (
 	"github.com/giantswarm/operatorkit/v6/pkg/resource/crud"
 	"github.com/giantswarm/operatorkit/v6/pkg/resource/k8s/configmapresource"
 	"github.com/giantswarm/operatorkit/v6/pkg/resource/k8s/secretresource"
+
 	"github.com/giantswarm/operatorkit/v6/pkg/resource/wrapper/metricsresource"
 	"github.com/giantswarm/operatorkit/v6/pkg/resource/wrapper/retryresource"
-	"github.com/giantswarm/resource/v3/appresource"
+	"github.com/giantswarm/resource/v4/appresource"
 	"k8s.io/apimachinery/pkg/labels"
-	apiv1alpha3 "sigs.k8s.io/cluster-api/api/v1alpha4"
+	capi "sigs.k8s.io/cluster-api/api/v1alpha4"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/giantswarm/cluster-apps-operator/pkg/project"
@@ -62,7 +63,7 @@ func NewCluster(config ClusterConfig) (*Cluster, error) {
 			K8sClient: config.K8sClient,
 			Logger:    config.Logger,
 			NewRuntimeObjectFunc: func() client.Object {
-				return new(apiv1alpha3.Cluster)
+				return new(capi.Cluster)
 			},
 			Resources: resources,